# api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/v1/images/{id}/media:
    get:
      tags:
      - image-rest-controller
      operationId: getImageFile
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    put:
      tags:
      - image-rest-controller
      operationId: replaceImageFile
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/garments/{id}:
    get:
      tags:
      - garment-rest-controller
      operationId: getGarment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GarmentExtendedDTO"
    put:
      tags:
      - garment-rest-controller
      operationId: replaceGarment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GarmentExtendedDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GarmentExtendedDTO"
    delete:
      tags:
      - garment-rest-controller
      operationId: deleteGarment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GarmentExtendedDTO"
...
  /api/v1/garments/{id}/images/:
    post:
      tags:
      - garment-rest-controller
      operationId: createGarmentImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ImageDTO"
...
  /api/v1/garments/:
    get:
      tags:
      - garment-rest-controller
      operationId: getGarments
      parameters:
      - name: nameSearch
        in: query
        required: false
        schema:
          type: string
      - name: categorySearch
        in: query
        required: false
        schema:
          type: string
      - name: minPrice
        in: query
        required: false
        schema:
          type: number
      - name: maxPrice
        in: query
        required: false
        schema:
          type: number
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelGarmentBasicDTO"
    post:
      tags:
      - garment-rest-controller
      operationId: createGarment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GarmentBasicDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GarmentExtendedDTO"
...
  /api/v1/images/{id}:
    get:
      tags:
      - image-rest-controller
      operationId: getImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ImageDTO"
...
  /api/v1/garments/{id}/images/{imageId}:
    delete:
      tags:
      - garment-rest-controller
      operationId: deleteGarmentImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: imageId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ImageDTO"
components:
  schemas:
    ImageDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
...
    GarmentBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: number
        image:
          $ref: "#/components/schemas/ImageDTO"
        reference:
          type: string
        category:
          type: string
        description:
          type: string
        features:
          type: string
...
    GarmentExtendedDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: number
        image:
          $ref: "#/components/schemas/ImageDTO"
        reference:
          type: string
        category:
          type: string
        description:
          type: string
        features:
          type: string
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
...
    PagedModelGarmentBasicDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/GarmentBasicDTO"
        page:
          $ref: "#/components/schemas/PageMetadata"