# backend/src/main/java/es/dawgrupo2/zendashop/controller/GarmentRestController.java
package es.dawgrupo2.zendashop.controller;

import java.io.IOException;
import java.math.BigDecimal;
import java.net.URI;
import java.util.List;
import java.util.NoSuchElementException;
import java.util.ArrayList;
import java.util.Optional;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Pageable;
import org.springframework.data.web.PageableDefault;
import org.springframework.data.domain.Page;
import org.springframework.http.ResponseEntity;
import org.springframework.security.access.AccessDeniedException;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.multipart.MultipartFile;

import es.dawgrupo2.zendashop.basicDTO.GarmentBasicDTO;
import es.dawgrupo2.zendashop.extendedDTO.GarmentExtendedDTO;
import es.dawgrupo2.zendashop.basicDTO.GarmentBasicMapper;
import es.dawgrupo2.zendashop.basicDTO.ImageDTO;
import es.dawgrupo2.zendashop.basicDTO.ImageMapper;
import es.dawgrupo2.zendashop.extendedDTO.GarmentExtendedMapper;
import es.dawgrupo2.zendashop.model.Garment;
import es.dawgrupo2.zendashop.model.User;
import es.dawgrupo2.zendashop.model.Image;
import es.dawgrupo2.zendashop.service.GarmentService;
import es.dawgrupo2.zendashop.service.UserService;
import es.dawgrupo2.zendashop.service.ImageService;
import jakarta.servlet.http.HttpServletRequest;

import static org.springframework.web.servlet.support.ServletUriComponentsBuilder.fromCurrentContextPath;
import static org.springframework.web.servlet.support.ServletUriComponentsBuilder.fromCurrentRequest;

@RestController
@RequestMapping("/api/v1/garments")
public class GarmentRestController {

	@Autowired
	private GarmentService garmentService;
	
    @Autowired
    private UserService userService;

    @Autowired
	private ImageService imageService;

	@Autowired
	private GarmentBasicMapper garmentBasicMapper;

    @Autowired
    private GarmentExtendedMapper garmentExtendedMapper;

    @Autowired
	private ImageMapper imageMapper;

    @GetMapping("/")
	public Page<GarmentBasicDTO> getGarments(@RequestParam(required = false) String nameSearch,
			@RequestParam(required = false) String categorySearch, @RequestParam(required = false) BigDecimal minPrice,
			@RequestParam(required = false) BigDecimal maxPrice,
			@PageableDefault(size = 10) Pageable pageable, HttpServletRequest request) {

        return garmentService.findAvailableGarmentsByOptionalFilters(nameSearch, categorySearch,
				minPrice, maxPrice, pageable).map(garmentBasicMapper::toDTO);
    }

	// If the user is logged in, show personalized offers
    @GetMapping("/offers")
    public List<GarmentBasicDTO> getOffers(HttpServletRequest request){
        List<Garment> garments = new ArrayList<Garment>();
        if (request.getUserPrincipal() != null) {
            String userEmail = request.getUserPrincipal().getName();
            Optional<User> userOpt = userService.findByEmail(userEmail);

            if (userOpt.isPresent()) {
                garments = garmentService.findSmartRecommendations(userOpt.get().getId());
            }
        } else {
            throw new AccessDeniedException("Tienes que estar loggeado para obtener recomendaciones");
        }
        return garmentBasicMapper.toDTOs(garments);
    }

	@GetMapping("/{id}")
	public GarmentExtendedDTO getGarment(@PathVariable long id) {
        return garmentExtendedMapper.toDTO(garmentService.findByIdAndAvailableTrue(id).orElseThrow(() -> new NoSuchElementException("Prenda no encontrada")));
	}

	@PostMapping("/")
	public ResponseEntity<GarmentExtendedDTO> createGarment(@RequestBody GarmentBasicDTO garmentBasicDTO){
		Garment garment = garmentBasicMapper.toDomain(garmentBasicDTO);
        
		String errorMsg = garmentService.validateFields(garment);
		if (errorMsg != null && !errorMsg.isEmpty()) {
			throw new IllegalArgumentException(errorMsg);
		}
		garment = garmentService.create(garment);
		GarmentExtendedDTO garmentDTO = garmentExtendedMapper.toDTO(garment);

		URI location = fromCurrentRequest().path("/{id}").buildAndExpand(garmentDTO.id()).toUri();

		return ResponseEntity.created(location).body(garmentDTO);
	}

    @PostMapping("/{id}/images/")
	public ResponseEntity<ImageDTO> createGarmentImage(@PathVariable long id, @RequestParam MultipartFile imageFile)
			throws IOException {
        
        Garment garment = garmentService.findByIdAndAvailableTrue(id).orElseThrow(() -> new NoSuchElementException("Prenda no encontrada"));

		if (garment.getImage() != null) {
			throw new IllegalStateException("La prenda ya tiene una imagen asignada, no se pueden asignar dos imágenes a una misma prenda, edítala o elimínala para cambiarla");
		}

		if (imageFile.isEmpty()) {
			throw new IllegalArgumentException("La imagen no puede estar vacía");
		}
		
        Image image = imageService.createImage(imageFile.getInputStream());
		garmentService.setImage(garment, image);
		
        URI location = fromCurrentContextPath()
				.path("/images/{imageId}/media")
				.buildAndExpand(image.getId())
				.toUri();

		return ResponseEntity.created(location).body(imageMapper.toDTO(image));
	}

    @PutMapping("/{id}")
	public GarmentExtendedDTO replaceGarment(@PathVariable long id, @RequestBody GarmentExtendedDTO updatedGarmentDTO) {
        Garment originalGarment = garmentService.findById(id).orElseThrow(() -> new NoSuchElementException("Prenda no encontrada"));
        Garment updateGarment = garmentExtendedMapper.toDomain(updatedGarmentDTO);

        String errorMsg = garmentService.validateFields(updateGarment);
		if (errorMsg != null && !errorMsg.isEmpty()) {
			throw new IllegalArgumentException(errorMsg);
		}
		return garmentExtendedMapper.toDTO(garmentService.updateGarment(originalGarment, updateGarment));
	}

    @DeleteMapping("/{id}")
	public GarmentExtendedDTO deleteGarment(@PathVariable long id) {
        Garment garment = garmentService.findByIdAndAvailableTrue(id).orElseThrow(() -> new NoSuchElementException("Prenda no encontrada"));
		return garmentExtendedMapper.toDTO(garmentService.disable(garment));
	}

    @DeleteMapping("/{id}/images/{imageId}")
	public ImageDTO deleteGarmentImage(@PathVariable long id, @PathVariable long imageId)
			throws IOException {
        Garment garment = garmentService.findByIdAndAvailableTrue(id).orElseThrow(() -> new NoSuchElementException("Prenda e imagen no encontradas"));
		if (garment.getImage() == null || garment.getImage().getId() != imageId) {
			throw new IllegalStateException("La imagen especificada no está asignada a esta prenda");
		}        
		ImageDTO image = imageMapper.toDTO(garment.getImage());
		garment.setImage(null);
		imageService.deleteImage(image.id());
		garmentService.save(garment);

		return image;
	}
}